+++ src/models/historial_curso_estudiante.rs
@@ -8,7 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = true)]
     pub id: i32,
     pub curso_id: i32,
-    pub estudiante_id: i64,
+    pub estudiante_id: i32,
     pub fecha_inscripcion: DateTime<Utc>,
     pub estado: String,
     pub calificacion_final: Option<f64>,
